+++ src/components/account/OpRegister.js
@@ -20,6 +20,8 @@ import {APIRequest} from '../.../../../network/Entity';
 import HorizontalInfoItem from './HorizontalInfoItem';
 import {CONFIG_KEY, isConfigActive} from '../../helper/configKeyHandler';
 import Loading from '../Loading';
+import {GPS_LIST_TYPE} from 'src/constants';
+import {servicesHandler, SERVICES_TYPE} from 'app-helper/servicesHandler';
 
 class OpRegister extends Component {
   constructor(props) {
@@ -40,7 +42,10 @@ class OpRegister extends Component {
         name: store.user_info ? store.user_info.city : '',
         id: store.user_info ? store.user_info.city_id : '',
       },
-      warehouseSelected: {id: store?.user_info?.store_id},
+      warehouseSelected: {
+        id: store?.user_info?.store_id,
+        name: store?.user_info?.store_name,
+      },
       licenseChecked: false,
       birth: store.user_info ? store.user_info.birth : '',
       cities: [],
@@ -73,7 +78,7 @@ class OpRegister extends Component {
 
   componentDidMount() {
     isConfigActive(CONFIG_KEY.CHOOSE_CITY_SITE_KEY) && this.getCities();
-    isConfigActive(CONFIG_KEY.CHOOSE_STORE_SITE_KEY) && this.getListWarehouse();
+    // isConfigActive(CONFIG_KEY.CHOOSE_STORE_SITE_KEY) && this.getListWarehouse();
     Actions.refresh({
       onBack: () => {
         this._unMount();
@@ -292,22 +297,30 @@ class OpRegister extends Component {
     this.setState({birth});
   };
 
-  onSelectWarehouse = (warehouseSelected, closeModal) => {
+  onSelectWarehouse = (warehouseSelected) => {
     this.setState({warehouseSelected});
-    closeModal();
   };
 
   onPressWarehouse = () => {
     Keyboard.dismiss();
-    Actions.push(appConfig.routes.modalList, {
-      heading: this.props.t('modal.store.title'),
-      data: this.state.listWarehouse,
-      selectedItem: this.state.warehouseSelected,
-      onPressItem: this.onSelectWarehouse,
-      onCloseModal: Actions.pop,
-      modalStyle: {
-        height: null,
-        maxHeight: '80%',
+    // Actions.push(appConfig.routes.modalList, {
+    //   heading: this.props.t('modal.store.title'),
+    //   data: this.state.listWarehouse,
+    //   selectedItem: this.state.warehouseSelected,
+    //   onPressItem: this.onSelectWarehouse,
+    //   onCloseModal: Actions.pop,
+    //   modalStyle: {
+    //     height: null,
+    //     maxHeight: '80%',
+    //   },
+    // });
+    servicesHandler({
+      type: SERVICES_TYPE.GPS_LIST_STORE,
+      selectedStore: this.state.warehouseSelected,
+      placeholder: this.props.t('gpsStore:searchSalePointPlaceholder'),
+      onPress: (store) => {
+        Actions.pop();
+        this.onSelectWarehouse(store);
       },
     });
   };
@@ -386,18 +399,18 @@ class OpRegister extends Component {
 
   renderWarehouse() {
     if (isConfigActive(CONFIG_KEY.CHOOSE_STORE_SITE_KEY)) {
-      const disable =
-        !this.state.listWarehouse || this.state.listWarehouse.length === 0;
+      const disable = false;
+      // !this.state.listWarehouse || this.state.listWarehouse.length === 0;
       const wareHouseData = {
         title: (
           <Text>
-            {this.props.t('data.chooseStore.title')}{' '}
+            {this.props.t('data.chooseSalePoint.title')}{' '}
             <Text style={styles.textRequired}>*</Text>
           </Text>
         ),
         value:
           this.state.warehouseSelected?.name ||
-          this.props.t('data.chooseStore.placeholder'),
+          this.props.t('data.chooseSalePoint.placeholder'),
         isLoading: this.state.isWarehouseLoading,
         select: true,
         disable,
@@ -806,4 +819,6 @@ const styles = StyleSheet.create({
   },
 });
 
-export default withTranslation(['opRegister', 'common'])(observer(OpRegister));
+export default withTranslation(['opRegister', 'common', 'gpsStore'])(
+  observer(OpRegister),
+);